requirements.txt
connexion[swagger-ui]==2.6.0
python-jose[cryptography]
SQLAlchemy==1.3.15
gunicorn==20.0.4
gevent==1.4.0
